requirements.txt
Flask==1.1.1
gunicorn==20.0.4
boto3==1.9.218
werkzeug==0.15.5
requests==2.22.0
pymongo==3.9.0
dnspython==1.16.0
pillow==6.2.1